requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==22.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.8.0
Django==4.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.10.0
future==0.18.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.8
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.5.0
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==1.3.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==2.0.2
openai==1.51.0
openai-whisper==20240930
openpyxl==3.1.5
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.4
PyPDF2==3.0.1
pypdfium2==4.30.0
pyTelegramBotAPI==4.22.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-pptx==1.0.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
Send2Trash==1.8.3
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.1
telebot==0.0.5
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.4.0
torch==2.5.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.13.1
yookassa==3.4.1
zipp==3.16.2
